Replaced tabs with spaces as requested by OE-Core migration guide
Python functions should always use four spaces for indentation and not tabs - a warning will be produced if tabs are found. Previously four spaces was the convention but many python functions used tabs. This has been fixed in OE-Core and other layers, and since python uses indentation for statement blocks, is very important that you make the same changes to python functions in your recipes.